4.23 exercises

+++ controllers/blogs.js
@@ -1,20 +1,26 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
 blogsRouter.get("/:id", async (request, response) => {
-  await Blog.findById(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id).populate("user", { username: 1, name: 1 });
+  response.json(blog);
 });
 
 blogsRouter.post("/", async (request, response) => {
-  const { title, author, url, likes, userId} = request.body;
-  const user = await User.findById(userId);
+  const { title, author, url, likes} = request.body;
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken || !decodedToken.id) {
+    return response.status(401).json({ error: "Token missing or invalid" });
+  }
+  const user = request.user;
 
   if (!title || !url) {
     return response.status(400).json({ error: "Title or URL is missing" });
@@ -37,6 +43,17 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const user = request.user;
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) {
+    return response.status(404).json({ error: "Blog not found" });
+  }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: "Unauthorized" });
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
